+++ pywren_ibm_cloud/compute/backends/knative/config.py
@@ -166,3 +166,6 @@ def load_config(config_data):
 
     if 'workers' not in config_data['pywren']:
         config_data['pywren']['workers'] = CONCURRENT_WORKERS_DEFAULT
+
+    if 'ibm_cos' in config_data and 'private_endpoint' in config_data['ibm_cos']:
+        del config_data['ibm_cos']['private_endpoint']